pyproject.toml
[project]
name = "ahc-template"
version = "0.5.0"
description = ""
authors = [{ name = "hitonanode", email = "32937551+hitonanode@users.noreply.github.com" }]
requires-python = ">=3.11,<3.12"
dependencies = [
    "numpy>=2.2.5,<3",
    "matplotlib>=3.10.1,<4",
    "pandas>=2.2.3,<3",
    "openpyxl>=3.1.5,<4",
    "pyyaml>=6.0.2,<7",
    "online-judge-tools>=11.5.1,<12",
    "optuna>=4.3.0,<5",
    "google-api-python-client>=2.167.0,<3",
    "gspread>=6.2.0,<7",
    "gspread-dataframe>=4.0.0,<5",
    "pandas-stubs>=2.2.3.250308,<3",
    "plotly>=6.0.1,<7",
    "pydantic>=2.11.3,<3",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0,<2",
    "ruff>=0.11.6,<0.12",
    "types-pyyaml>=6.0.12.20250402,<7",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = "3.11"

[tool.ruff]
target-version = "py311"
lint.select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I001"
]
